+++ mybooking-templates/mybooking-plugin-reservation-tmpl.php
@@ -955,6 +955,6 @@
         </button>
       </div>
     </div>
-  </div>
+  </form>
 
 </script>